refactor(swingset): split security flow harnesses by section
